Extract AccessLog into a separate module (#3336)
